requirements.txt
transformers
torch
gtts
pandas
streamlit
numpy
pillow